+++ sb3_contrib/common/maskable/__init__.py
@@ -1,1 +1,0 @@
-"""Maskable shims."""