+++ sb3_contrib/common/maskable/policies.py
@@ -1,5 +1,0 @@
-"""Maskable policy shims."""
-
-
-class MaskableActorCriticPolicy:
-    pass