+++ utils/results_utils.py
@@ -1,3 +1,0 @@
-"""Compatibility shim for legacy `utils.results_utils` imports."""
-
-from ztb.utils.results_utils import *  # noqa: F401,F403