+++ ztb/config/schemas/__init__.py
@@ -1,7 +1,0 @@
-"""
-Configuration schemas.
-"""
-
-from ztb.config.schemas.zaif import GlobalConfig, ZaifTradeBotConfig
-
-__all__ = ["GlobalConfig", "ZaifTradeBotConfig"]