+++ ztb/ops/monitoring/__init__.py
@@ -1,1 +1,0 @@
-# ztb.ops.monitoring package