+++ ztb/trading/environment/heavy_env/__init__.py
@@ -1,8 +1,0 @@
-"""Heavy trading environment package."""
-
-from .core import FlipHeavyTradingEnv, HeavyTradingEnv
-
-__all__ = [
-    "HeavyTradingEnv",
-    "FlipHeavyTradingEnv",
-]