+++ ztb/training/evaluation/__init__.py
@@ -1,5 +1,0 @@
-# ztb.training.evaluation package
-
-from .eval_gates import EvalGates
-
-__all__ = ["EvalGates"]